requirement.txt
flask == 3.1.0 
python-dotenv == 1.1.0
psycopg2-binary == 2.9.10
bcrypt == 4.3.0